Move legacy stuff into separate plugins so that logic can easily be removed later
