+++ build.gradle
@@ -57,13 +57,9 @@ gradlePlugin {
             id = "eu.xenit.docker-config"
             implementationClass = "eu.xenit.gradle.docker.config.DockerConfigPlugin"
         }
-        dockerCore {
-            id = "eu.xenit.docker.core"
-            implementationClass = "eu.xenit.gradle.docker.core.DockerPlugin"
-        }
         docker {
             id = "eu.xenit.docker"
-            implementationClass = "eu.xenit.gradle.docker.DockerPlugin"
+            implementationClass = "eu.xenit.gradle.docker.core.DockerPlugin"
         }
         dockerCompose {
             id = "eu.xenit.docker-compose"
@@ -88,10 +84,6 @@ pluginBundle {
 
         dockerConfig {
             displayName = "Docker configuration plugin"
-        }
-
-        dockerCore {
-            displayName = "Docker plugin"
         }
 
         docker {
