+++ src/main/java/eu/xenit/gradle/docker/DockerLegacyPlugin.java
@@ -1,20 +1,19 @@
 package eu.xenit.gradle.docker;
 
-import eu.xenit.gradle.docker.autotag.DockerAutotagPlugin;
-import eu.xenit.gradle.docker.config.DockerConfigPlugin;
 import eu.xenit.gradle.docker.core.DockerExtension;
 import eu.xenit.gradle.docker.internal.Deprecation;
 import eu.xenit.gradle.docker.tasks.internal.DockerBuildImage;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 
-public class DockerPlugin implements Plugin<Project> {
-    public static final String PLUGIN_ID = "eu.xenit.docker";
+/**
+ * @deprecated Use {@link eu.xenit.gradle.docker.core.DockerPlugin} instead. Will be removed in 6.0
+ */
+@Deprecated
+public class DockerLegacyPlugin implements Plugin<Project> {
 
     @Override
     public void apply(Project project) {
-        project.getPluginManager().apply(eu.xenit.gradle.docker.core.DockerPlugin.class);
-
         DockerExtension dockerExtension = project.getExtensions().getByType(DockerExtension.class);
 
         DockerFileExtension dockerFileExtension = project.getExtensions()
