+++ src/test/java/eu/xenit/gradle/docker/compose/DockerComposeConventionImplTest.java
@@ -6,7 +6,7 @@ import static org.junit.Assert.assertTrue;
 import com.avast.gradle.dockercompose.ComposeSettings;
 import com.bmuschko.gradle.docker.tasks.image.DockerBuildImage;
 import eu.xenit.gradle.docker.alfresco.DockerAlfrescoPlugin;
-import eu.xenit.gradle.docker.DockerPlugin;
+import eu.xenit.gradle.docker.core.DockerPlugin;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
